Change size() function to return a size_t so that result size is not inadvertently truncated.
BUG=None
TEST=Ran existing unit tests and trybots.

Review URL: http://codereview.chromium.org/6386008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@73001 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/pickle.h
@@ -55,8 +55,7 @@ class Pickle {
   Pickle& operator=(const Pickle& other);
 
   // Returns the size of the Pickle's data.
-  int size() const { return static_cast<int>(header_size_ +
-                                             header_->payload_size); }
+  size_t size() const { return header_size_ + header_->payload_size; }
 
   // Returns the data for this Pickle.
   const void* data() const { return header_; }
